Make the frontpage the rules list when logged in.
