+++ frontend/src/components/PageHeader.vue
@@ -1,11 +1,6 @@
 <script setup lang="ts">
 import { CContainer, CNavbar, CNavbarNav } from "@coreui/bootstrap-vue";
-import { useI18n } from "vue-i18n";
-import { useUserStore } from "../stores/user";
 import LoginButton from "./LoginButton.vue";
-import MainNavLink from "./MainNavLink.vue";
-const userStore = useUserStore();
-const { t } = useI18n();
 </script>
 
 <template>
@@ -19,9 +14,6 @@ const { t } = useI18n();
         <img src="@/assets/logo.png" alt="FMN logo" height="40" />
       </router-link>
       <CNavbarNav class="align-items-center">
-        <MainNavLink to="/rules" v-if="userStore.loggedIn">{{
-          t("rules")
-        }}</MainNavLink>
         <LoginButton />
       </CNavbarNav>
     </CContainer>
