+++ frontend/src/views/HomeView.vue
@@ -1,20 +1,37 @@
 <script setup lang="ts">
-const api_docs_url = import.meta.env.VITE_API_URL + "/docs";
+import { useRulesQuery } from "@/api/rules";
+import { CAlert, CSpinner } from "@coreui/bootstrap-vue";
+import RulesList from "../components/RulesList.vue";
+import { useUserStore } from "../stores/user";
+
+const userStore = useUserStore();
+
+// Getting rules
+const { isLoading, isError, data, error } = useRulesQuery();
 </script>
 
 <template>
-  <div class="p-5 mb-4">
-    <div class="home">
-      <h1 class="fw-bold display-5">Notifications</h1>
-      <p class="col-md-8 fs-4">
-        Centrally managed preferences for Fedora Infrastructure notifications to
-        your inbox, irc client, and mobile device.
-      </p>
+  <template v-if="userStore.username">
+    <div class="rules">
+      <h1>My Rules</h1>
+      <div v-if="isLoading" class="text-center"><CSpinner /></div>
+      <CAlert v-else-if="isError" color="danger"
+        >Could not load the rules: {{ error }}</CAlert
+      >
+      <RulesList v-else-if="data" :rules="data" />
     </div>
-  </div>
-  <div class="mt-5">
-    <p>View online API docs <a :href="api_docs_url">here</a>.</p>
-  </div>
+  </template>
+  <template v-else>
+    <div class="p-5 mb-4">
+      <div class="home">
+        <h1 class="fw-bold display-5">Notifications</h1>
+        <p class="col-md-8 fs-4">
+          Centrally managed preferences for Fedora Infrastructure notifications
+          to your inbox, irc client, and mobile device.
+        </p>
+      </div>
+    </div>
+  </template>
 </template>
 
 <style></style>
